Move DB utils to db.py file
